Refactor di GestoreDashboardView.vue
